Rename _common.py to common.py to expose public API
Public API components (ContainerError, Dependency) should not be in a
module with underscore prefix. Renamed _common.py to common.py to make
it a proper public module.

- Updated all imports from ._common to .common
- All 61 tests passing
- All linters passing with 10.00/10 pylint rating
