+++ .venv2/bin/python
@@ -1,0 +1,1 @@
+python3
