+++ .venv2/bin/python3
@@ -1,0 +1,1 @@
+/usr/local/bin/python3
