+++ .venv2/bin/python3.11
@@ -1,0 +1,1 @@
+python3
