+++ .venv2/bin/trove-classifiers
@@ -1,0 +1,8 @@
+#!/home/user/independency/.venv2/bin/python3
+# -*- coding: utf-8 -*-
+import re
+import sys
+from trove_classifiers.__main__ import cli
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(cli())
