+++ .venv2/pyvenv.cfg
@@ -1,0 +1,5 @@
+home = /usr/local/bin
+include-system-site-packages = false
+version = 3.11.14
+executable = /usr/bin/python3.11
+command = /usr/local/bin/python3 -m venv /home/user/independency/.venv2
