+++ independency/__init__.py
@@ -1,4 +1,4 @@
 # flake8: noqa
-from ._common import ContainerError, Dependency
+from .common import ContainerError, Dependency
 from .async_container import AsyncContainer, AsyncContainerBuilder
 from .container import Container, ContainerBuilder
