+++ independency/async_container.py
@@ -1,7 +1,7 @@
 import asyncio
 from typing import Any, Callable, Dict, Set
 
-from ._common import (
+from .common import (
     ObjType,
     Registration,
     build_container,
