+++ tests/test_async_container.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, Final, Generic, TypeVar
 import pytest
 
 from independency import ContainerError, Dependency as Dep
-from independency._common import get_generic_mapping, get_signature
+from independency.common import get_generic_mapping, get_signature
 from independency.async_container import AsyncContainer, AsyncContainerBuilder
